Change expiresIn's type from int to long

+++ trip-kit-booking/src/main/java/com/skedgo/android/tripkit/booking/AccessTokenResponse.java
@@ -13,7 +13,7 @@ public interface AccessTokenResponse {
 
   @SerializedName("access_token") String accessToken();
   @SerializedName("token_type") String tokenType();
-  @SerializedName("expires_in") int expiresIn();
+  @SerializedName("expires_in") long expiresIn();
   @Nullable @SerializedName("refresh_token") String refreshToken();
 
 }
